refactor base yoshi config
