--- test/fixtures/javascript/base-template/baseConfig.js
@@ -1,7 +1,0 @@
-module.exports = {
-  projectType: 'app',
-  externals: {
-    react: 'React',
-    'react-dom': 'ReactDOM',
-  },
-};
