+++ test/javascript/features/externals/yoshi.config.js
@@ -1,4 +1,4 @@
-const config = require('./baseConfig');
+const config = require('../../../../test/fixtures/javascript/base-template/yoshi.config');
 
 config.externals = { ...config.externals, externals: 'externals' };
 module.exports = config;
